Refactor tests to use Quick
